feat: add search and loader
