+++ package.json
@@ -34,7 +34,7 @@
     "vite": "^5.3.1"
   },
   "lint-staged": {
-    "**/*": [
+    "**/*.tsx": [
       "eslint --fix",
       "prettier --write --ignore-unknown"
     ]
