+++ src/components/bottomLine.tsx
@@ -1,35 +1,10 @@
 import { Component, ReactNode } from 'react';
 import PeopleItem from './peopleItem';
+import { state, props } from '../types/types';
 
 const BASE_PATH = 'https://swapi.dev/api';
-const SEARC_PATH = '/people';
-export type props = {
-  birth_year: string;
-  eye_color: string;
-  films?: string[];
-  gender: string;
-  hair_color: string;
-  height: string;
-  homeworld?: string;
-  mass: string;
-  name: string;
-  skin_color: string;
-};
-
-type state = {
-  result: {
-    birth_year: string;
-    eye_color: string;
-    films?: string[];
-    gender: string;
-    hair_color: string;
-    height: string;
-    homeworld?: string;
-    mass: string;
-    name: string;
-    skin_color: string;
-  }[];
-};
+const PAGE_PATH = '/people';
+const SEARCH_PATH = '/?search=';
 
 class BottomLine extends Component {
   constructor(props: props) {
@@ -37,43 +12,72 @@ class BottomLine extends Component {
   }
 
   state: state = {
+    searchQuerry: '',
     result: [],
+    loading: false,
   };
 
   componentDidMount(): void {
-    // const {searchQuerry} = this.state
-    fetch(`${BASE_PATH}${SEARC_PATH}`)
+    const { searchQuerry } = this.state;
+    this.fetchData(searchQuerry);
+  }
+
+  fetchData = (searchQuerry: string) => {
+    this.setState({
+      loading: true,
+    });
+    fetch(`${BASE_PATH}${PAGE_PATH}${SEARCH_PATH}${searchQuerry}`)
       .then((res) => res.json())
       .then((result) => {
         this.setResult(result.results);
+        this.setState({
+          loading: false,
+        });
       })
       .catch((error) => error);
-  }
+  };
+
+  handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    this.setState({
+      searchQuerry: e.target.value,
+    });
+  };
+
+  getSearch = () => {
+    const { searchQuerry } = this.state;
+    this.fetchData(searchQuerry);
+  };
 
   setResult = (result: Response) => {
     this.setState({ result });
-    console.log(result);
   };
   render(): ReactNode {
-    const { result } = this.state;
+    const { searchQuerry, result, loading } = this.state;
 
     return (
-      <div className="bottomLine">
-        <>
-          {result.map((item: props) => (
-            <PeopleItem
-              birth_year={item.birth_year}
-              eye_color={item.eye_color}
-              gender={item.gender}
-              hair_color={item.hair_color}
-              height={item.height}
-              mass={item.mass}
-              name={item.name}
-              skin_color={item.skin_color}
-            />
-          ))}
-        </>
-      </div>
+      <>
+        <div className="topLine">
+          <input type="search" onChange={this.handleInputChange} value={searchQuerry} />
+          <button onClick={this.getSearch}>Search</button>
+        </div>
+        {loading ? <div>Загрузка</div> : null}
+        <div className="bottomLine">
+          <>
+            {result.map((item: props) => (
+              <PeopleItem
+                birth_year={item.birth_year}
+                eye_color={item.eye_color}
+                gender={item.gender}
+                hair_color={item.hair_color}
+                height={item.height}
+                mass={item.mass}
+                name={item.name}
+                skin_color={item.skin_color}
+              />
+            ))}
+          </>
+        </div>
+      </>
     );
   }
 }
