+++ src/types/types.tsx
@@ -1,0 +1,29 @@
+export type props = {
+  birth_year: string;
+  eye_color: string;
+  films?: string[];
+  gender: string;
+  hair_color: string;
+  height: string;
+  homeworld?: string;
+  mass: string;
+  name: string;
+  skin_color: string;
+};
+
+export type state = {
+  searchQuerry: string;
+  result: {
+    birth_year: string;
+    eye_color: string;
+    films?: string[];
+    gender: string;
+    hair_color: string;
+    height: string;
+    homeworld?: string;
+    mass: string;
+    name: string;
+    skin_color: string;
+  }[];
+  loading: boolean;
+};
